pyproject.toml
[tool.poetry]
name = "qa-guru-python-07-20"
version = "0.1.0"
description = ""
authors = ["Anhelina Babareka <anhielina.babareka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.4.2"
requests = "^2.31.0"
jsonschema = "^4.19.1"
allure-pytest = "^2.13.2"
curlify = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
